+++ src/commands/cat/cat.js
