+++ src/commands/cp/cp.js
