+++ src/commands/rm/rm.js
